joining_tables
TASK: Write a query to find the top 10 countries for Rockbuster in terms of customer numbers. 
Creating the code: 
First, I reviewed the ERD for the database to determine where to find the information I am looking for. Then, I decided to refer to the 4 tables using A,B,C,D. 
In SELECT, I determine what do I want to see as result of this query. 
In FROM, I started with the table with the first piece of information: the count of customer: Table A is “customer”; Column to be count is “customer_id”. 
In INNER JOIN, I start with the first table that has the link to table A: Table B is “address” and is linked by the “address_id”. 
The next INNER JOIN is the table connected to Table B: Table C is “city” and is linked by the “city_id”.  
The next INNER JOIN is the table connected to Table C: Table D is “country”, and is linked by the “country_id” (This table contain one of the desired result in SELECT: country_names). 
Next, I made sure that customer count is aggregated per country, hence GROUP BY D.country. Then, to see countries with the largest number of customer first I applied ORDER BY COUNT(A.customer_id) DESC. Last, I applied LIMIT 10 for top 10.